+++ python-libraries/data-platform-catalogue/data_platform_catalogue/entities.py
@@ -98,8 +98,6 @@ class DataProductMetadata:
 
 class SecurityClassification(Enum):
     OFFICIAL = auto()
-    SECRET = auto()
-    TOP_SECRET = auto()
 
 
 @dataclass